Subdomain for the ML Course changed to course18
This is my super tiny first change, but the link for the ML course used to be https://course.fast.ai/ml and it got moved to https://course18.fast.ai/ml. There's a page that redirects people to the new page, but people could save a click by just changing the URL here.